+++ app/Util/ActivityPub/Inbox.php
@@ -37,6 +37,7 @@ use App\Util\ActivityPub\Validator\UndoFollow as UndoFollowValidator;
 
 use App\Services\PollService;
 use App\Services\FollowerService;
+use App\Services\ReblogService;
 use App\Services\StatusService;
 use App\Services\UserFilterService;
 use App\Services\NetworkTimelineService;
@@ -569,13 +570,9 @@ class Inbox
 			return;
 		}
 
-		if(Helpers::validateLocalUrl($activity) == false) {
-			return;
-		}
-
 		$parent = Helpers::statusFetch($activity);
 
-		if(empty($parent)) {
+		if(!$parent || empty($parent)) {
 			return;
 		}
 
@@ -590,18 +587,23 @@ class Inbox
 			'type' => 'share'
 		]);
 
-		Notification::firstOrCreate([
-			'profile_id' => $parent->profile->id,
-			'actor_id' => $actor->id,
-			'action' => 'share',
-			'message' => $status->replyToText(),
-			'rendered' => $status->replyToHtml(),
-			'item_id' => $parent->id,
-			'item_type' => 'App\Status'
-		]);
+		Notification::firstOrCreate(
+			[
+				'profile_id' => $parent->profile_id,
+				'actor_id' => $actor->id,
+				'action' => 'share',
+				'item_id' => $parent->id,
+				'item_type' => 'App\Status',
+			], [
+				'message' => $status->replyToText(),
+				'rendered' => $status->replyToHtml(),
+			]
+		);
 
 		$parent->reblogs_count = $parent->reblogs_count + 1;
 		$parent->save();
+
+		ReblogService::addPostReblog($parent->profile_id, $status->id);
 
 		return;
 	}
@@ -790,17 +792,23 @@ class Inbox
 				if(is_array($obj) && isset($obj['object'])) {
 					$obj = $obj['object'];
 				}
-				if(!is_string($obj) || !Helpers::validateLocalUrl($obj)) {
+				if(!is_string($obj)) {
 					return;
 				}
-				$status = Status::whereUri($obj)->exists();
+				if(Helpers::validateLocalUrl($obj)) {
+					$parsedId = last(explode('/', $obj));
+					$status = Status::find($parsedId);
+				} else {
+					$status = Status::whereUri($obj)->first();
+				}
 				if(!$status) {
 					return;
 				}
 				Status::whereProfileId($profile->id)
 					->whereReblogOfId($status->id)
-					->forceDelete();
-				Notification::whereProfileId($status->profile->id)
+					->delete();
+				ReblogService::removePostReblog($profile->id, $status->id);
+				Notification::whereProfileId($status->profile_id)
 					->whereActorId($profile->id)
 					->whereAction('share')
 					->whereItemId($status->reblog_of_id)